app_engine/requirements.txt
Flask==0.12.2
google-cloud-vision==0.27.0
gunicorn==19.7.1
